Massive Refactoring: split code in TwoDayViewPresenter across several classes
