+++ optashift-employee-rostering-gwtui/src/main/java/org/optaplanner/openshift/employeerostering/gwtui/client/calendar/TimeRowDrawableProvider.java
@@ -1,5 +1,6 @@
 package org.optaplanner.openshift.employeerostering.gwtui.client.calendar;
 
+import org.optaplanner.openshift.employeerostering.gwtui.client.calendar.twodayview.TwoDayViewPresenter;
 import org.optaplanner.openshift.employeerostering.gwtui.client.interfaces.HasTimeslot;
 
 public interface TimeRowDrawableProvider<G extends HasTitle, I extends HasTimeslot<G>, T extends TimeRowDrawable<G>> {
