+++ optashift-employee-rostering-gwtui/src/main/resources/org/optaplanner/openshift/employeerostering/gwtui/client/calendar/twodayview/TwoDayView.html
